requirements.txt
pyperclip
python-decouple
streamlit
openai
python-docx
fpdf
PyPDF2
chardet
